YandexART/requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-SQLAlchemy>=2.5
SQLAlchemy>=1.4
requests>=2.28
python-dotenv>=0.21
gunicorn>=21.0 